Rename to AssertIssetToSpecificMethodRector
