+++ tests/Rector/Contrib/PHPUnit/SpecificMethod/AssertIssetToSpecificMethodRector/Correct/correct.php.inc
